styling frontend

+++ db/schema.rb
@@ -15,7 +15,7 @@ ActiveRecord::Schema[7.1].define(version: 2023_11_14_023259) do
   enable_extension "plpgsql"
 
   create_table "messages", force: :cascade do |t|
-    t.integer "room_id", null: false
+    t.bigint "room_id", null: false
     t.text "message"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
